requirements.txt
openai
pandas
numpy
faiss-cpu
PyMuPDF
python-dotenv
